Changelog 0.2.0-beta

+++ pkg/debrid/structs/realdebrid.go
@@ -74,7 +74,7 @@ type RealDebridTorrentInfo struct {
 	Filename         string `json:"filename"`
 	OriginalFilename string `json:"original_filename"`
 	Hash             string `json:"hash"`
-	Bytes            int    `json:"bytes"`
+	Bytes            int64  `json:"bytes"`
 	OriginalBytes    int    `json:"original_bytes"`
 	Host             string `json:"host"`
 	Split            int    `json:"split"`
